Centralize and deduplicate Rollup configs
