+++ packages/liveblocks-react-ui/package.json
@@ -96,6 +96,7 @@
   "devDependencies": {
     "@liveblocks/eslint-config": "*",
     "@liveblocks/jest-config": "*",
+    "@liveblocks/rollup-config": "*",
     "@rollup/plugin-replace": "^5.0.5",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.1.1",
@@ -104,19 +105,6 @@
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "msw": "^0.27.1",
-    "postcss": "^8.4.31",
-    "postcss-advanced-variables": "^3.0.1",
-    "postcss-combine-duplicated-selectors": "^10.0.3",
-    "postcss-functions": "^4.0.2",
-    "postcss-import": "^15.1.0",
-    "postcss-lightningcss": "^1.0.0",
-    "postcss-nesting": "^12.0.1",
-    "postcss-reporter": "^7.0.5",
-    "postcss-sort-media-queries": "^5.2.0",
-    "rollup": "^3.28.0",
-    "rollup-plugin-dts": "^5.3.1",
-    "rollup-plugin-esbuild": "^5.0.0",
-    "rollup-plugin-preserve-directives": "^0.2.0",
     "stylelint": "^15.10.2",
     "stylelint-config-standard": "^34.0.0",
     "stylelint-order": "^6.0.3",
